cdk/cms-cdk/requirements-dev.txt
pytest==6.2.5
boto3
awscdk_resources_mongodbatlas
aws-cdk-lib
constructs>=10.0.0,<11.0.0
python-dotenv

